+++ test/recording_studio_notifications_push_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class RecordingStudioNotificationsPushTest < Minitest::Test
   def test_version_matches_release
-    assert_equal "0.1.5", ::RecordingStudioNotificationsPush::VERSION
+    assert_equal "0.1.6", ::RecordingStudioNotificationsPush::VERSION
   end
 
   def test_engine_exists
@@ -98,10 +98,13 @@ class RecordingStudioNotificationsPushTest < Minitest::Test
     assert_includes devices_show, "push_enable: true"
     assert_includes devices_show, "Push Notifications"
     assert_includes devices_show, "Get notifications on your devices"
+    assert_includes devices_show, "Manage notifications"
+    assert_includes devices_show, "@notifications_settings_path"
     assert_includes devices_show, "FlatPack::List::Component"
     assert_includes devices_show, "installation.list_icon"
     assert_includes devices_show, 'icon: "trash"'
     assert_includes devices_show, "icon_only: true"
+    assert_includes devices_show, "turbo_method: :delete"
     refute_includes devices_show, "Active browsers and devices"
     refute_includes devices_show, "No devices yet"
     refute_includes devices_show, "push_disable"